examples: swap the HTTP KV demo for a lean heartbeat node
Replace k8s_kv (which embedded a hand-rolled HTTP server — too much
incidental machinery for a library example) with k8s_heartbeat: each pod
periodically writes its own heartbeat/<pod> key and a pre-insert hook logs
each key on first sight. Since the hook fires on peer-originated updates too,
reconciliation between pods is visible directly in kubectl logs, with no
extra port, API, or dependency.

Reverts the tokio dev-dep and Dockerfile EXPOSE additions; kind now builds
the k8s_heartbeat example and the README walkthrough is log-based.
