+++ Dockerfile
@@ -9,8 +9,9 @@
 # The container reads its configuration from the environment — see deploy/k8s/.
 #
 # Which example to compile is selectable via the EXAMPLE build arg (default `k8s_node`, the
-# production node). The local kind playground builds `k8s_kv`, which adds a demo HTTP key/value API:
-#   docker build --build-arg EXAMPLE=k8s_kv -t reconcile:kind .
+# production node). The local kind playground builds `k8s_heartbeat`, which additionally writes a
+# per-pod heartbeat key and logs reconciled keys so convergence is visible in the logs:
+#   docker build --build-arg EXAMPLE=k8s_heartbeat -t reconcile:kind .
 
 # Selectable at build time; redeclared inside each stage that uses it (Docker ARG scoping).
 ARG EXAMPLE=k8s_node
@@ -35,10 +36,8 @@ RUN cargo build --release --example "${EXAMPLE}" --features metrics-prometheus \
 FROM gcr.io/distroless/cc-debian12 AS runtime
 COPY --from=builder /usr/local/bin/reconcile-node /usr/local/bin/reconcile-node
 
-# Gossip (UDP), metrics/probes (TCP), and the optional demo HTTP KV API (TCP, k8s_kv only).
-# Match deploy/k8s/ (and deploy/kind/) and the container env.
+# Gossip (UDP) and metrics/probes (TCP). Match deploy/k8s/ (and deploy/kind/) and the container env.
 EXPOSE 8080/udp
-EXPOSE 8081/tcp
 EXPOSE 9000/tcp
 
 # Run as the distroless non-root user.
